Refactor : Basic Code Refactoring #1
1. Admin, Guest Controller 변경
2. Domain Method 요청 URI 주소 변경
3. Package 변경
