+++ src/main/java/com/solver/solver_be/domain/user/controller/AdminController.java
@@ -1,8 +1,7 @@
 package com.solver.solver_be.domain.user.controller;
 
-import com.solver.solver_be.domain.user.dto.*;
+import com.solver.solver_be.domain.user.dto.requestDto.*;
 import com.solver.solver_be.domain.user.service.AdminService;
-import com.solver.solver_be.domain.user.service.GuestService;
 import com.solver.solver_be.global.response.GlobalResponseDto;
 import com.solver.solver_be.global.security.webSecurity.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
@@ -18,77 +17,45 @@ import javax.validation.Valid;
 @Slf4j
 @RestController
 @RequiredArgsConstructor
-public class UserController {
+public class AdminController {
 
     private final AdminService adminService;
-    private final GuestService guestService;
 
     // 1. Admin SignUp
-    @PostMapping("/signup/business")
+    @PostMapping("/admin/signup")
     public ResponseEntity<GlobalResponseDto> signupBusiness(@Valid @RequestBody AdminSignupRequestDto signupRequestDto) {
         return adminService.signupBusiness(signupRequestDto);
     }
 
     // 2. Admin Login
-    @PostMapping("/login/business")
+    @PostMapping("/admin/login")
     public ResponseEntity<GlobalResponseDto> loginBusiness(@RequestBody LoginRequestDto loginRequestDto, HttpServletResponse response) {
         return adminService.login(loginRequestDto, response);
     }
 
-    // 4. Admin change password
-    @PutMapping("/change/password/business")
+    // 3. Change Admin Password
+    @PutMapping("/admin/password")
     public ResponseEntity<GlobalResponseDto> changeBusiness(@Valid @RequestBody PasswordChangeRequestDto passwordChangeRequestDto,
                                                             @AuthenticationPrincipal UserDetailsImpl userDetails) {
         return adminService.changePassword(passwordChangeRequestDto, userDetails.getAdmin());
     }
 
-    // 5. Found Admin userId
-    @PostMapping("/admin/search")
+    // 4. Find Admin UserId
+    @PostMapping("/admin/user-id")
     public ResponseEntity<GlobalResponseDto> adminIdSearch(@RequestBody UserSearchRequestDto userSearchRequestDto) throws MessagingException {
         return adminService.findAdminSearchId(userSearchRequestDto);
     }
 
-    // 6. Reset Admin Password
+    // 5. Reset Admin Password
     @PostMapping("/admin/password")
     public ResponseEntity<GlobalResponseDto> resetAdminPassword(@RequestBody PasswordResetRequestDto passwordResetRequestDto) throws MessagingException {
         return adminService.resetAdminPassword(passwordResetRequestDto);
     }
 
-    // 7. Guest Signup
-    @PostMapping("/signup/guest")
-    public ResponseEntity<GlobalResponseDto> signupGuest(@Valid @RequestBody GuestSignupRequestDto signupRequestDto) {
-        return guestService.signupGuest(signupRequestDto);
-    }
-
-    // 8. Guest Login
-    @PostMapping("/login/guest")
-    public ResponseEntity<GlobalResponseDto> loginGuest(@RequestBody LoginRequestDto loginRequestDto, HttpServletResponse response) {
-        return guestService.login(loginRequestDto, response);
-    }
-
-    // 9. Guest change password
-    @PutMapping("/change/password/guest")
-    public ResponseEntity<GlobalResponseDto> changeGuest(@Valid @RequestBody PasswordChangeRequestDto passwordChangeRequestDto,
-                                                         @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        return guestService.changePassword(passwordChangeRequestDto, userDetails.getGuest());
-    }
-
-    // 10. Found Guest userId
-    @PostMapping("/guest/search")
-    public ResponseEntity<GlobalResponseDto> guestIdSearch(@RequestBody UserSearchRequestDto userSearchRequestDto) throws MessagingException {
-        return guestService.findGuestSearchId(userSearchRequestDto);
-    }
-
-    // 11. Reset Guest Password
-    @PostMapping("/guest/password")
-    public ResponseEntity<GlobalResponseDto> resetGuestPassword(@RequestBody PasswordResetRequestDto passwordResetRequestDto) throws MessagingException {
-        return guestService.resetGuestPassword(passwordResetRequestDto);
-    }
-
-    // 12. Create Lobby Id
-    @PostMapping("/signup/lobbyId")
+    // 6. Create Admin LobbyId
+    @PostMapping("/admin/lobby")
     public ResponseEntity<GlobalResponseDto> createLobbyId(@RequestBody LobbyRequestDto lobbyRequestDto,
                                                            @AuthenticationPrincipal UserDetailsImpl userDetails) throws MessagingException {
-        return adminService.createLobbyId(lobbyRequestDto,userDetails.getAdmin());
+        return adminService.createLobbyId(lobbyRequestDto, userDetails.getAdmin());
     }
 }
