+++ src/main/java/com/solver/solver_be/domain/user/dto/requestDto/LobbyRequestDto.java
@@ -1,4 +1,4 @@
-package com.solver.solver_be.domain.user.dto;
+package com.solver.solver_be.domain.user.dto.requestDto;
 
 import lombok.Getter;
