+++ src/main/java/com/solver/solver_be/domain/user/dto/requestDto/LoginRequestDto.java
@@ -1,9 +1,10 @@
-package com.solver.solver_be.domain.user.dto;
+package com.solver.solver_be.domain.user.dto.requestDto;
 
 import lombok.Getter;
 
 @Getter
 public class LoginRequestDto {
+
     private String userId;          // Guest/Admin Id
     private String password;        // Guest/Admin Password
 }
