+++ src/main/java/com/solver/solver_be/domain/user/dto/requestDto/PasswordChangeRequestDto.java
@@ -1,4 +1,4 @@
-package com.solver.solver_be.domain.user.dto;
+package com.solver.solver_be.domain.user.dto.requestDto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,9 +9,11 @@ import javax.validation.constraints.Pattern;
 @Getter
 @NoArgsConstructor
 public class PasswordChangeRequestDto {
+
     private String password;
 
     @NotBlank(message = "비밀번호는 필수입니다.")
     @Pattern(regexp = "^[a-zA-Z0-9~!@#$%^&*()_+=?,./<>{}\\[\\]\\-]{8,15}$", message = "비밀번호는 8~15자리 영문 대소문자(a~z, A~Z), 숫자(0~9), 특수문자를 사용하세요.")
     private String newPassword;
+
 }
