+++ src/main/java/com/solver/solver_be/domain/user/service/AdminService.java
@@ -2,7 +2,8 @@ package com.solver.solver_be.domain.user.service;
 
 import com.solver.solver_be.domain.company.entity.Company;
 import com.solver.solver_be.domain.company.repository.CompanyRepository;
-import com.solver.solver_be.domain.user.dto.*;
+import com.solver.solver_be.domain.user.dto.requestDto.*;
+import com.solver.solver_be.domain.user.dto.responseDto.LoginResponseDto;
 import com.solver.solver_be.domain.user.entity.Admin;
 import com.solver.solver_be.domain.user.entity.Guest;
 import com.solver.solver_be.domain.user.repository.AdminRepository;
@@ -27,13 +28,12 @@ import org.springframework.transaction.annotation.Transactional;
 import javax.mail.MessagingException;
 import javax.servlet.http.HttpServletResponse;
 import java.util.Optional;
-import java.util.Random;
 
 @Service
 @RequiredArgsConstructor
 public class AdminService {
-
     private final JwtUtil jwtUtil;
+    private final InfoProvider infoProvider;
     private final EmailService emailService;
     private final PasswordEncoder passwordEncoder;
     private final AdminRepository adminRepository;
@@ -160,7 +160,7 @@ public class AdminService {
         if (emailService.verifyEmailCode(passwordResetRequestDto.getEmail(), passwordResetRequestDto.getCode())) {
 
             // Provisional Password Issue
-            String newPassword = generateRandomPassword();
+            String newPassword = infoProvider.generateRandomPassword();
 
             // Replace existing password
             admin.setPassword(passwordEncoder.encode(newPassword));
@@ -187,8 +187,8 @@ public class AdminService {
         }
 
         // Lobby ID Data
-        String userId = generateRandomId();
-        String password = generateRandomPassword();
+        String userId = infoProvider.generateRandomId();
+        String password = infoProvider.generateRandomPassword();
         String name = company.getCompanyName() + "LobbyId";
         String phoneNum = company.getCompanyCallNum();
         String companyToken = company.getCompanyToken();
@@ -215,33 +215,5 @@ public class AdminService {
         companyRepository.save(company);
 
         return ResponseEntity.ok(GlobalResponseDto.of(SuccessType.LOBBYID_SIGN_UP));
-    }
-
-    // Method : Generating a temporary password
-    private String generateRandomPassword() {
-        int leftLimit = 48; // number '0'
-        int rightLimit = 122; // alphabet 'z'
-        int targetStringLength = 10;
-        Random random = new Random();
-
-        return random.ints(leftLimit, rightLimit + 1)
-                .filter(i -> (i <= 57 || i >= 65) && (i <= 90 || i >= 97))
-                .limit(targetStringLength)
-                .collect(StringBuilder::new, StringBuilder::appendCodePoint, StringBuilder::append)
-                .toString();
-    }
-
-    // Method : Generating a temporary id
-    private String generateRandomId() {
-        int leftLimit = 48;                     // number '0'
-        int rightLimit = 122;                   // alphabet 'z'
-        int targetStringLength = 6;
-        Random random = new Random();
-
-        return random.ints(leftLimit, rightLimit + 1)
-                .filter(i -> (i <= 57 || i >= 65) && (i <= 90 || i >= 97))
-                .limit(targetStringLength)
-                .collect(StringBuilder::new, StringBuilder::appendCodePoint, StringBuilder::append)
-                .toString();
     }
 }
