+++ src/main/java/com/solver/solver_be/domain/visitform/controller/VisitFormController.java
@@ -66,11 +66,4 @@ public class VisitFormController {
                                                               @AuthenticationPrincipal UserDetailsImpl userDetails) {
         return visitorService.searchVisitForms(visitFormSearchRequestDto, userDetails.getAdmin());
     }
-    // 6. Sort VisitForms
-    @GetMapping("/visit-forms/sort")
-    public ResponseEntity<GlobalResponseDto> getLocation(@AuthenticationPrincipal UserDetailsImpl userDetails,
-                                                         @RequestParam(required = false, value = "orderBy") String orderBy,
-                                                         @RequestParam(value = "isAsc") Boolean isAsc) {
-        return visitorService.sortVisitForms(userDetails.getAdmin(), orderBy, isAsc);
-    }
 }
